examples/requirements.txt
ggtag
ggwave
pyaudio
pyserial
pypng
